+++ pdfalyzer/pdfalyzer.py
@@ -98,7 +98,7 @@ class Pdfalyzer:
         self.font_info_extraction_error: Optional[Exception] = None
         self.max_generation = 0  # PDF revisions are "generations"; this is the max generation encountered
         self.nodes_encountered: Dict[int, PdfTreeNode] = {}  # Nodes we've seen already
-        self._tree_nodes_by_id: Dict[int, PdfTreeNode | None] = {}  # Nodes in the tree we've looked up already
+        self._tree_nodes: Dict[int, PdfTreeNode | None] = {}  # Nodes in the tree we've looked up already
         self._indeterminate_ids = set()  # See INDETERMINATE_REF_KEYS comment
 
         # Bootstrap the root of the tree with the trailer. PDFs are always read trailer first.
@@ -147,8 +147,8 @@ class Pdfalyzer:
 
     def find_node_by_idnum(self, idnum: int) -> Optional[PdfTreeNode]:
         """Find node with `idnum` in the tree. Return `None` if that node is not reachable from the root."""
-        self._tree_nodes_by_id[idnum] = self._tree_nodes_by_id.get(idnum, self.find_node_with_attr('idnum', idnum, True))
-        return self._tree_nodes_by_id[idnum]
+        self._tree_nodes[idnum] = self._tree_nodes.get(idnum, self.find_node_with_attr('idnum', idnum, True))
+        return self._tree_nodes[idnum]
 
     def find_node_with_attr(self, attr: str, value: str | int, should_raise: bool = False) -> PdfTreeNode | None:
         """Find node with a property where you only expect one of those nodes to exist"""